+++ modules/config/LocationBlock/LocBlock.hpp
@@ -1,7 +1,7 @@
 #ifndef LOCBLOCK_HPP
 # define LOCBLOCK_HPP
 
-# include "../Tools/ParseTool.hpp"
+# include "../../Tools/ParseTool.hpp"
 # include "../otherBlock/OtherBlock.hpp"
 # include "../interface/IBlock.hpp"
 # include "../httpBlock/HttpBase.hpp"
