+++ modules/config/interface/IBlock.hpp
@@ -1,6 +1,6 @@
 #ifndef IBLOCK_HPP
 # define IBLOCK_HPP
-# include "../Tools/ParseTool.hpp"
+# include "../../Tools/ParseTool.hpp"
 class IBlock
 {
 public:
